chore!: reorder operands of the `==`
It is considered bad practice to have conditions like: `3 == x`, which
is what was previously generated. This also has semantic meaning, as
types that implement `PartialEq` are not always symmetric, so users will
expect that `x == 3` is generated when writing a struct with asymmetric
equality.

BREAKING CHANGE: Rust equality is not always symmetric, so structs that
have asymmetric `PartialEq` implementations will break.